Drop support for Play v2.7
The most recent release of Play 2.7 was v2.7.9 on 14th December 2020, over
3 years ago. Given that Play 2.8 is End Of Life (EOL) on 31st May 2024,
we should _definitely_ not still be using Play 2.7!

Removing Play 2.7 support allows cleaning up some deprecated methods
and cleans up build.sbt too.

* https://github.com/playframework/playframework/releases/tag/2.7.9
* https://github.com/playframework/playframework/releases/tag/2.8.21
