+++ .tool-versions
@@ -1,0 +1,1 @@
+java corretto-11.0.22.7.1
